refactor: keep heartbeat architecture debt flat
